Add configurable httpd timeout
This patch adds an apiTimeout field to the OctaviaSpecCore
to the already OctaviaSpec one to allow configure the timeouts
for HAProxy and Apache. Inherit from the controller in case
it is not specified by the user.

Also fixing *core* API spec version of the new
GetDefaultRouteAnnotations function.

Jira: https://issues.redhat.com/browse/OSPRH-6713